MODIFY: 회원 서비스 클래스 역할 분리
주소 관련 서비스와 회원 관련 API 서비스 클래스 분리
